backend/requirements/base.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.0
django-filter==23.3
django-reversion==5.0.4
django-celery-beat==2.5.0
channels==4.0.0
channels-redis==4.1.0
drf-yasg==1.21.7
psycopg2-binary==2.9.9
Pillow==10.1.0
celery==5.3.4
redis==5.0.1
boto3==1.28.63
jinja2==3.1.2
pyyaml==6.0.1
python-dotenv==1.0.0
gunicorn==21.2.0
requests==2.31.0
